financial-data-pipeline/data-analysis/requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
statsmodels>=0.13.0
psycopg2-binary>=2.9.0
cassandra-driver>=3.25.0
cachetools>=5.0.0
python-dotenv>=0.19.0
flask-cors==3.0.10
requests>=2.26.0
yfinance==0.2.54

